backend/requirements.txt
# ── Core API ──────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── ML / NLP ──────────────────────────────────────────────────
torch==2.2.2          # CPU-only; swap for torch+cu121 for GPU
transformers==4.40.1
sentencepiece==0.2.0
huggingface-hub==0.23.0

# ── Data / Evaluation ─────────────────────────────────────────
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.9.0
seaborn==0.13.2
pandas==2.2.2

# ── HTTP / Scraping ───────────────────────────────────────────
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1

# ── Config / Utils ────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
